abstracted confirm out to a component then used dynamic import to page fix build error
